+++ examples/examples.go
@@ -1,0 +1,6 @@
+// Package examples contains several example Go applications that use Lattigo in both the single- and multiparty settings,
+// as well as several example parameter sets. See examples/README.md for more information about the examples.
+//
+// Note that the code in this package, including the example parameter sets, is solely meant to illustrate the use of the library and facilitate quick experiments.
+// It should not be depended upon and may at any time be changed or be removed.
+package examples